Attempt of kick and role edit
There was apparently a bug in poking, which is causing the bot not to respond to the poke. However, it seems to be from the role which was on a different server. I am unsure what role that should belong to, and so I will go ahead and push what I have then let Star or Enterprise insert the correct role. 

I've also included the first version of the kick. Please do not use the kick command, as it will cause the bot to puke out errors.